+++ tests/utils_tests.go
@@ -3,7 +3,7 @@
 // Use of this source code is governed by a MIT-style license that
 // can be found in the LICENSE file for the project.
 
-package finviz
+package tests
 
 /*
 func TestExportScreenCSV(t *testing.T) {
